Move refCountToAlpha into its own file
