+++ packages/css/src/fileScope.ts
@@ -1,4 +1,5 @@
 import { onEndFileScope } from './adapter';
+import { refCountToAlpha } from './refCountToAlpha';
 import type { FileScope } from './types';
 
 let refCounter = 0;
@@ -29,22 +30,7 @@ export function getFileScope(): FileScope {
   return fileScopes[0];
 }
 
-const startingCharCode = 'a'.charCodeAt(0);
-
-export function refCountToAlpha(number: number) {
-  let n = number + 1;
-  let returnValue = '';
-
-  while (n > 0) {
-    const offset = (n - 1) % 26;
-    returnValue = String.fromCharCode(startingCharCode + offset) + returnValue;
-    n = ((n - offset) / 26) | 0;
-  }
-
-  return returnValue;
-}
-
-export function getAndIncrementRefCounter() {
+export function getAndIncrementRefCount() {
   const alpha = refCountToAlpha(refCounter);
   refCounter++;
